+++ app/cells/components/story_slider_cell.rb
@@ -1,5 +1,9 @@
 class Components::StorySliderCell < Cell::ViewModel
+  include StatisticsHelper
 
   private
 
+  def common_info
+    model || nil
+  end
 end